cleanup: move retry loops to DatabaseAdminConnection
