+++ google/cloud/spanner/internal/database_admin_logging.h
@@ -44,9 +44,6 @@ class DatabaseAdminLogging : public DatabaseAdminStub {
       google::spanner::admin::database::v1::CreateDatabaseRequest const&
           request) override;
 
-  future<StatusOr<google::spanner::admin::database::v1::Database>>
-      AwaitCreateDatabase(google::longrunning::Operation) override;
-
   StatusOr<google::spanner::admin::database::v1::Database> GetDatabase(
       grpc::ClientContext&,
       google::spanner::admin::database::v1::GetDatabaseRequest const&) override;
@@ -61,10 +58,6 @@ class DatabaseAdminLogging : public DatabaseAdminStub {
       grpc::ClientContext& context,
       google::spanner::admin::database::v1::UpdateDatabaseDdlRequest const&
           request) override;
-
-  future<
-      StatusOr<google::spanner::admin::database::v1::UpdateDatabaseDdlMetadata>>
-  AwaitUpdateDatabase(google::longrunning::Operation operation) override;
 
   Status DropDatabase(
       grpc::ClientContext& context,
