+++ google/cloud/spanner/internal/database_admin_metadata.h
@@ -40,9 +40,6 @@ class DatabaseAdminMetadata : public DatabaseAdminStub {
       google::spanner::admin::database::v1::CreateDatabaseRequest const&
           request) override;
 
-  future<StatusOr<google::spanner::admin::database::v1::Database>>
-      AwaitCreateDatabase(google::longrunning::Operation) override;
-
   StatusOr<google::spanner::admin::database::v1::Database> GetDatabase(
       grpc::ClientContext&,
       google::spanner::admin::database::v1::GetDatabaseRequest const&) override;
@@ -57,10 +54,6 @@ class DatabaseAdminMetadata : public DatabaseAdminStub {
       grpc::ClientContext& context,
       google::spanner::admin::database::v1::UpdateDatabaseDdlRequest const&
           request) override;
-
-  future<
-      StatusOr<google::spanner::admin::database::v1::UpdateDatabaseDdlMetadata>>
-  AwaitUpdateDatabase(google::longrunning::Operation operation) override;
 
   Status DropDatabase(
       grpc::ClientContext& context,
