+++ google/cloud/spanner/spanner_client.bzl
@@ -37,7 +37,6 @@ spanner_client_hdrs = [
     "internal/connection_impl.h",
     "internal/database_admin_logging.h",
     "internal/database_admin_metadata.h",
-    "internal/database_admin_retry.h",
     "internal/database_admin_stub.h",
     "internal/date.h",
     "internal/instance_admin_stub.h",
@@ -90,7 +89,6 @@ spanner_client_srcs = [
     "internal/connection_impl.cc",
     "internal/database_admin_logging.cc",
     "internal/database_admin_metadata.cc",
-    "internal/database_admin_retry.cc",
     "internal/database_admin_stub.cc",
     "internal/date.cc",
     "internal/instance_admin_stub.cc",
